// src/utils/enums/tipo.comprobante.enum.ts
/**
* @value Cfdi V4.0 - c_TipoDeComprobante
* */
export enum TipoComprobanteEnum {
    /**
    * @value Ingreso
    * */
    I = 'I',
    /**
    * @value Egreso
    * */
    E = 'E',
    /**
    * @value Traslado
    * */
    T = 'T',
    /**
    * @value Nómina
    * */
    N = 'N',
    /**
    * @value PagosPago
    * */
    P = 'P',
}